[Controls] Use EUI Selectable for Field search (#151231)
## Summary
Replaces Control field selection list with EUISelectable.

(cherry picked from commit eb9cc11a7c81073a75fb3e6ec33e61796cc40aff)

# Conflicts:
#	src/plugins/controls/public/control_group/editor/control_editor.tsx
