+++ src/plugins/presentation_util/public/components/field_picker/field_type_filter.scss
@@ -1,0 +1,6 @@
+.presFilterByType__panel {
+  min-width: $euiSizeXXL * 8;
+  @include euiBreakpoint('l', 'xl') {
+    width: $euiFormMaxWidth;
+  }
+}
